Try to show the marker of Meradin more prominently
Atm the "sicherer Ort" overlays the city "Meradin" - as it is part of it.
Adopting the coordinates a bit let the red marker of the city be seen a bit more.
Changing the order in the geo json file, however, didn't result in showing the
red marker above the blue one.